requirements.txt
# Web Framework
fastapi==0.135.1
uvicorn==0.15.0
python-socketio==5.10.0
aiohttp==3.9.1
python-multipart==0.0.6

# Database
# MongoDB (optional - only needed if DB_PROVIDER=mongodb)
motor==3.3.2
pymongo==4.6.1
# Redis (optional - for caching)
redis==5.0.1

# AI/ML Services
openai==1.10.0
groq==0.4.2
requests==2.31.0

# Audio Processing
# Note: We use FFmpeg directly via subprocess (no pydub needed)
# Note: openai-whisper is NOT needed - we use whisper-cli (C++ binary) instead
soundfile==0.12.1
numpy==1.26.3

# Audio Emotion Detection (OPTIONAL - requires Python 3.11-3.12)
# PyTorch is not yet compatible with Python 3.14
# To enable audio emotion detection:
# 1. Use Python 3.12 or 3.11, OR
# 2. Wait for PyTorch to release Python 3.14 wheels
# 
# Uncomment these if using Python 3.11-3.12:
# transformers<5.0
# torch>=2.2.0,<2.3.0
# torchaudio>=2.2.0,<2.3.0
# numpy<2.0
# protobuf
# Note: librosa is NOT needed - we use torchaudio for audio loading

# Utilities
python-dotenv==1.0.0
pydantic==2.12.5
loguru==0.7.2
python-dateutil==2.8.2
aiofiles==23.2.1
httpx==0.26.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0
